service manage api location change
